+++ eurocropsml/acquisition/region.py
@@ -7,6 +7,7 @@ from typing import cast
 import geopandas as gpd
 import pandas as pd
 import pyogrio
+from pyogrio.errors import DataSourceError
 from tqdm import tqdm
 
 from eurocropsml.acquisition.config import CollectorConfig
@@ -73,7 +74,8 @@ def add_nuts_regions(
     try:
         nuts: gpd.GeoDataFrame = pyogrio.read_dataframe(nuts_regions_file)
 
-    except FileNotFoundError:
+    except DataSourceError:
+
         available_crs: list[str] = []
         available_years: list[int] = []
         for file in nuts_dir.iterdir():